Dedup contactIds at render time
We have some live workspace documents with `contactIds` arrays that have
duplicate identity IDs. This causes `Encountered two children with the same
key` warnings in `Share`. The patch fixes those warnings by deduping the
array before rendering.

We could and probably eventually should fix this in the data layer by
converting `contactIds` to a map (i.e. a set), which would structurally
avoid the race condition. We've chosen to fix for now in the view because:

* There aren't any other apparent issues caused by these duplicate IDs
* It's easier and less risky
* It's reversible, more so than data migrations