uniquelySelected no longer required on CME
We can no longer require the `uniquelySelected` prop on `CodeMirrorEditor`
because the editor may be displayed at the top-level in workspace (e.g. if a
text card is double-clicked).

For now just make the warning go away. Eventually we may want to rethink
selection / uniq selection / focus and its interaction with which document
is top-level. But another day.