Don't re-render every card every time we drag one
Currently we re-`render` every `Card` every time the `Board` changes, which
includes any time we drag a card 1 pixel. With boards like our current one
with ~50 cards this causes a lot of churn. This change adjusts what we pass
for `dragState` to avoid generating a new object for every card for every board
render in the common case that there is no drag state for that card. Just use
null and handle it in the card.