Put together inspector view using observablehq
- Inspector document at the moment just contains `targetURL` field of the
document being inspected, but in the future it will be extended to store
state like (un)folded substructures and possibly more.
- At the moment view initially renders title editor to enter `targetURL`
but ideally it would possibly have dropdown like omnibar to just
choose document from the active board. I also would like to have
long press on Option/Alt key to activate overlay over widgets with a
button to launch an inspector for any of the widgets.